#1338 - Factor CAS <-> brat conversion code into Pojos
- Moved conversion code from BratWriter to DKPro2Brat
